allow NavList to paginate multiple groups
