+++ app/components/primer/beta/nav_list/group.html.erb
@@ -1,7 +1,9 @@
-<% with_post_list_content do %>
-  <% if show_more_item? %>
-    <%= show_more_item %>
+<nav-list-group>
+  <% with_post_list_content do %>
+    <% if show_more_item? %>
+      <%= show_more_item %>
+    <% end %>
   <% end %>
-<% end %>
 
-<% render_parent %>
+  <% render_parent %>
+</nav-list-group>
