+++ demo/app/controllers/nav_list_items_controller.rb
@@ -5,14 +5,19 @@
 class NavListItemsController < ApplicationController
   layout false
 
-  PAGES = {
-    2 => [
-      { label: "Bachelor Chow", href: "/foods/bachelor-chow" },
-      { label: "LöBrau", href: "/foods/lobrau" }
-    ]
-  }.freeze
+  FOODS = [
+    { label: "Bachelor Chow", href: "/foods/bachelor-chow" },
+    { label: "LöBrau", href: "/foods/lobrau" },
+    { label: "Taco Bellevue Hospital", href: "/foods/taco-bellevue-hospital" },
+    { label: "Olde Fortran", href: "/foods/olde-fortran" },
+    { label: "Space Honey", href: "/foods/space-honey" },
+    { label: "Spice Weasel", href: "/foods/spice-weasel" },
+  ]
 
   def index
-    @data = PAGES[params[:page].to_i]
+    items_per_page = 2
+    # the first page is already shown in the nav list, so we need to offset the starting index
+    start_index = (params[:page].to_i - items_per_page) * items_per_page
+    @data = FOODS.slice(start_index, items_per_page)
   end
 end
